+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Adjust Rustdoc's json parsing for format v26.
+  See https://github.com/rust-lang/rust/pull/111427 for the breaking change.
+
 ## v0.3.2
 
 - Fix bad release in v0.3.1